fixed the date format
